+++ src/components/User/KycRequests/KycRequest.List.vue
@@ -43,7 +43,7 @@
       </div>
 
       <div class="request-list__list-wrp">
-        <template v-if="list.data && list.data.length">
+        <template v-if="list && list.length">
           <div class="app-list">
             <div class="app-list__header">
               <span class="app-list__cell">Email</span>
@@ -53,7 +53,7 @@
             </div>
             <button
               class="app-list__li"
-              v-for="item in list.data"
+              v-for="item in list"
               :key="item.id"
               @click="toggleViewMode(item.requestor.id)"
             >
@@ -73,17 +73,6 @@
               </span>
             </button>
           </div>
-
-          <div class="app__more-btn-wrp">
-            <button
-              class="app__btn-secondary"
-              v-if="!isListEnded && list.data"
-              @click="onMoreClick"
-            >
-              More
-            </button>
-          </div>
-
         </template>
 
         <template v-else>
@@ -106,6 +95,14 @@
           </div>
         </template>
 
+        <div class="app__more-btn-wrp">
+          <collection-loader
+            :first-page-loader="loadList"
+            @first-page-load="setList"
+            @next-page-load="extendList"
+            ref="collectionLoaderBtn"
+          />
+        </div>
       </div>
     </template>
 
@@ -135,6 +132,7 @@ import config from '@/config'
 import { ApiCallerFactory } from '@/api-caller-factory'
 import { clearObject } from '@/utils/clearObject'
 import { ErrorHandler } from '@/utils/ErrorHandler'
+import { CollectionLoader } from '@/components/common'
 
 const VIEW_MODES_VERBOSE = Object.freeze({
   index: 'index',
@@ -143,12 +141,20 @@ const VIEW_MODES_VERBOSE = Object.freeze({
 
 export default {
   name: 'kyc-request-list',
-  components: { UserView, EmailGetter, SelectField, InputField },
+  components: { UserView, EmailGetter, SelectField, InputField, CollectionLoader },
+  provide () {
+    const kycRequestsList = {}
+    Object.defineProperty(kycRequestsList, 'updateAsk', {
+      enumerable: true,
+      get: () => false,
+      set: () => this.loadList()
+    })
+    return { kycRequestsList }
+  },
   data () {
     return {
       isLoading: false,
-      isListEnded: false,
-      list: {},
+      list: [],
       view: {
         mode: VIEW_MODES_VERBOSE.index,
         userId: null,
@@ -177,14 +183,11 @@ export default {
     }
   },
 
-  created () {
-    this.loadList()
-  },
-
   methods: {
     snakeToCamelCase,
     async loadList () {
       this.isLoading = true
+      let response = {}
       try {
         let address = ''
         if (this.filters.address) {
@@ -192,7 +195,7 @@ export default {
         } else if (this.filters.email) {
           address = await this.getAccountIdByEmail()
         }
-        this.list = await ApiCallerFactory
+        response = await ApiCallerFactory
           .createCallerInstance()
           .getWithSignature('/v3/change_role_requests', {
             filter: clearObject({
@@ -202,23 +205,19 @@ export default {
             }),
             include: ['request_details.account']
           })
-        this.isListEnded = !(this.list.data || []).length
       } catch (error) {
-        ErrorHandler.process(error)
+        ErrorHandler.processWithoutFeedback(error)
       }
       this.isLoading = false
+      return response
     },
 
-    async onMoreClick () {
-      try {
-        const oldLength = this.list.data.length
-        const chunk = await this.list.fetchNext()
-        this.list._data = this.list.data.concat(chunk.data)
-        this.list.fetchNext = chunk.fetchNext
-        this.isListEnded = oldLength === this.list.data.length
-      } catch (error) {
-        ErrorHandler.process(error)
-      }
+    setList (data) {
+      this.list = data
+    },
+
+    async extendList (data) {
+      this.list = this.list.concat(data)
     },
 
     toggleViewMode (id) {
@@ -236,29 +235,17 @@ export default {
       })
     },
 
-    async getAccountIdByEmail () {
-      let address
-      try {
-        const { data } = await ApiCallerFactory
-          .createCallerInstance()
-          .getWithSignature('/identities', {
-            filter: clearObject({
-              email: this.filters.email
-            })
-          })
-        address = data[0].address
-      } catch (e) {
-        address = ''
-      }
-      return address
+    reloadCollectionLoader () {
+      this.$refs.collectionLoaderBtn.loadFirstPage()
     },
     formatDate
   },
   watch: {
-    'filters.state' () { this.loadList() },
-    'filters.roleToSet' () { this.loadList() },
-    'filters.email': _.throttle(function () { this.loadList() }, 1000),
-    'filters.address': _.throttle(function () { this.loadList() }, 1000)
+    'filters.state' () { this.reloadCollectionLoader() },
+    'filters.roleToSet' () { this.reloadCollectionLoader() },
+    'filters.address': _.throttle(function () {
+      this.reloadCollectionLoader()
+    }, 1000)
   }
 }
 </script>